refactor: split default commit messages and default pr titles file
