+++ lib/default-commit-messages.js
@@ -12,11 +12,6 @@ const defaultCommitMessages = {
   // Conditionally appended to dependencyUpdate
   closes: '\n\nCloses #${number}'
 }
-
-const defaultPRTitles = {
-  basicPR: 'Update ${dependency} to the latest version 🚀',
-  groupPR: 'Update ${dependency} in group ${group} to the latest version 🚀'
-}
 /* eslint-enable no-template-curly-in-string */
 
-module.exports = { defaultCommitMessages, defaultPRTitles }
+module.exports = { defaultCommitMessages }
