+++ lib/default-pr-titles.js
@@ -1,0 +1,8 @@
+/* eslint-disable no-template-curly-in-string */
+const defaultPrTitles = {
+  basicPR: 'Update ${dependency} to the latest version 🚀',
+  groupPR: 'Update ${dependency} in group ${group} to the latest version 🚀'
+}
+/* eslint-enable no-template-curly-in-string */
+
+module.exports = { defaultPrTitles }
