+++ lib/get-config.js
@@ -1,6 +1,7 @@
 const _ = require('lodash')
 const mergejson = require('mergejson')
-const { defaultCommitMessages, defaultPRTitles } = require('./default-config-messages')
+const { defaultCommitMessages } = require('./default-commit-messages')
+const { defaultPRTitles } = require('./default-pr-titles')
 
 module.exports = repository => {
   const greenkeeperConfig = _.get(repository, 'greenkeeper', {})
