+++ lib/get-message.js
@@ -1,4 +1,5 @@
-const { defaultCommitMessages, defaultPRTitles } = require('./default-config-messages')
+const { defaultCommitMessages } = require('./default-commit-messages')
+const { defaultPRTitles } = require('./default-pr-titles')
 
 function replaceVariables (message, variables) {
   let replacedString = message
